+++ packages/react-data-grid-addons/src/cells/headerCells/filters/index.ts
@@ -1,4 +1,0 @@
-export { default as NumericFilter } from './NumericFilter';
-export { default as AutoCompleteFilter } from './AutoCompleteFilter';
-export { default as MultiSelectFilter } from './MultiSelectFilter';
-export { default as SingleSelectFilter } from './SingleSelectFilter';