+++ CHANGELOG.md
@@ -15,10 +15,14 @@
       - ⚠️ This replace the `rowGetter` and `rowsCount` props
   - `column.cellClass(row)` function support:
     - `column = { ..., cellClass(row) { return string; } }`
+  - `column.minWidth`
+  - `column.maxWidth`
   - `column.headerCellClass`
   - `scrollToRow` method
     - ⚠️ This replaces the `scrollToRowIndex` prop
 - **Removed:**
+  - **Support:**
+    - ⚠️ IE11
   - **Packages:**
     - ⚠️ `react-data-grid-addons`
   - **Props:**